requirements.txt
asgiref==3.2.10
Babel==2.8.0
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
decorator==4.4.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.8
django-braces==1.14.0
django-cors-headers==3.4.0
django-countries==6.1.2
django-environ==0.4.5
django-filter==2.3.0
django-modeltranslation==0.15.1
django-money==1.1
django-oauth-toolkit==1.3.2
django-phonenumber-field==4.0.0
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
ipython==7.16.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
parso==0.7.1
phonenumbers==8.12.7
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
py-moneyed==0.8.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rest-framework-generic-relations==2.0.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.1.0
